backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary>=2.9.9
sqlalchemy==2.0.23
python-jose==3.3.0
cryptography>=41.0.0
bcrypt>=4.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
python-dotenv==1.0.0
weasyprint==62.3
pydyf==0.10.0
jinja2==3.1.2
Pillow>=10.4.0
asyncpg
reportlab>=4.0.0
aiofiles
openpyxl
PyPDF2==3.0.1